Intent-Simulator/requirements.txt
flask==3.0.2
flask-cors==4.0.0
rdflib==7.0.0
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
openai==1.12.0 